+++ Sources/Views/InputTextView.swift
@@ -49,6 +49,17 @@ open class InputTextView: UITextView {
         }
     }
     
+    /// The images that are currently stored as NSTextAttachment's
+    open var images: [UIImage] {
+        return parseForAttachedImages()
+    }
+    
+    open var components: [Any] {
+        return parseForComponents()
+    }
+    
+    open var isImagePasteEnabled: Bool = true
+    
     /// A UILabel that holds the InputTextView's placeholder text
     open let placeholderLabel: UILabel = {
         let label = UILabel()
@@ -148,7 +159,9 @@ open class InputTextView: UITextView {
         layer.cornerRadius = 5.0
         layer.borderWidth = 1.25
         layer.borderColor = UIColor.lightGray.cgColor
+        allowsEditingTextAttributes = false
         setupPlaceholderLabel()
+        setupObservers()
     }
     
     /// Adds the placeholderLabel to the view and sets up its initial constraints
@@ -168,6 +181,14 @@ open class InputTextView: UITextView {
         placeholderLabelConstraintSet?.activate()
     }
     
+    /// Adds the required notification observers
+    private func setupObservers() {
+        
+        NotificationCenter.default.addObserver(self,
+                                               selector: #selector(InputTextView.redrawTextAttachments),
+                                               name: .UIDeviceOrientationDidChange, object: nil)
+    }
+    
     /// Updates the placeholderLabels constraint constants to match the placeholderLabelInsets
     private func updateConstraintsForPlaceholderLabel() {
         
@@ -176,4 +197,143 @@ open class InputTextView: UITextView {
         placeholderLabelConstraintSet?.left?.constant = placeholderLabelInsets.left
         placeholderLabelConstraintSet?.right?.constant = -placeholderLabelInsets.right
     }
+    
+    // MARK: - Image Paste Support
+    
+    open override func canPerformAction(_ action: Selector, withSender sender: Any?) -> Bool {
+        
+        if action == NSSelectorFromString("paste:") && UIPasteboard.general.image != nil {
+            return isImagePasteEnabled
+        }
+        return super.canPerformAction(action, withSender: sender)
+    }
+    
+    open override func paste(_ sender: Any?) {
+        
+        guard let image = UIPasteboard.general.image else {
+            return super.paste(sender)
+        }
+        pasteImageInTextContainer(with: image)
+    }
+    
+    /// Addes a new UIImage to the NSTextContainer as an NSTextAttachment
+    ///
+    /// - Parameter image: The image to add
+    private func pasteImageInTextContainer(with image: UIImage) {
+
+        // Add the new image as an NSTextAttachment
+        let attributedImageString = NSAttributedString(attachment: textAttachment(using: image))
+        
+        let isEmpty = attributedText.length == 0
+        
+        // Add a new line character before the image, this is what iMessage does
+        let newAttributedStingComponent = isEmpty ? NSMutableAttributedString(string: "") : NSMutableAttributedString(string: "\n")
+        newAttributedStingComponent.append(attributedImageString)
+        
+        // Add a new line character after the image, this is what iMessage does
+        newAttributedStingComponent.append(NSAttributedString(string: "\n"))
+        
+        // The attributes that should be applied to the new NSAttributedString to match the current attributes
+        let attributes: [NSAttributedStringKey:Any] = [
+            NSAttributedStringKey.font : font ?? UIFont.preferredFont(forTextStyle: .body),
+            NSAttributedStringKey.foregroundColor : textColor ?? .black,
+            ]
+        newAttributedStingComponent.addAttributes(attributes, range: NSRange(location: 0, length: newAttributedStingComponent.length))
+        
+        textStorage.beginEditing()
+        // Paste over selected text
+        textStorage.replaceCharacters(in: selectedRange, with: newAttributedStingComponent)
+        textStorage.endEditing()
+        
+        // Advance the range to the selected range plus the number of characters added
+        let location = selectedRange.location + (isEmpty ? 2 : 3)
+        selectedRange = NSRange(location: location, length: 0)
+    
+        // Broadcast a notification to recievers such as the MessageInputBar which will handle resizing
+        NotificationCenter.default.post(name: .UITextViewTextDidChange, object: self)
+    }
+    
+    /// Returns an NSTextAttachment the provided image that will fit inside the NSTextContainer
+    ///
+    /// - Parameter image: The image to create an attachment with
+    /// - Returns: The formatted NSTextAttachment
+    private func textAttachment(using image: UIImage) -> NSTextAttachment {
+        
+        guard let cgImage = image.cgImage else { return NSTextAttachment() }
+        let scale = image.size.width / (frame.width - 2 * (textContainerInset.left + textContainerInset.right))
+        let textAttachment = NSTextAttachment()
+        textAttachment.image = UIImage(cgImage: cgImage, scale: scale, orientation: .up)
+        return textAttachment
+    }
+    
+    /// Returns all images that exist as NSTextAttachment's
+    ///
+    /// - Returns: An array of type UIImage
+    private func parseForAttachedImages() -> [UIImage] {
+        
+        var images = [UIImage]()
+        let range = NSRange(location: 0, length: attributedText.length)
+        attributedText.enumerateAttribute(.attachment, in: range, options: [], using: { value, range, stop -> Void in
+            
+            if let attachment = value as? NSTextAttachment {
+                if let image = attachment.image {
+                    images.append(image)
+                } else if let image = attachment.image(forBounds: attachment.bounds,
+                                                       textContainer: nil,
+                                                       characterIndex: range.location) {
+                    images.append(image)
+                }
+            }
+        })
+        return images
+    }
+    
+    /// Returns an array of components (either a String or UIImage) that makes up the textContainer in
+    /// the order that they were typed
+    ///
+    /// - Returns: An array of objects guaranteed to be of UIImage or String
+    private func parseForComponents() -> [Any] {
+        
+        var components = [Any]()
+        let range = NSRange(location: 0, length: attributedText.length)
+        attributedText.enumerateAttributes(in: range, options: []) { (object, range, stop) in
+            
+            if object.keys.contains(.attachment){
+                if let attachment = object[.attachment] as? NSTextAttachment {
+                    if let image = attachment.image {
+                        components.append(image)
+                    } else if let image = attachment.image(forBounds: attachment.bounds,
+                                                           textContainer: nil,
+                                                           characterIndex: range.location) {
+                        components.append(image)
+                    }
+                }
+            } else {
+                let stringValue = attributedText.attributedSubstring(from: range).string.trimmingCharacters(in: .whitespacesAndNewlines)
+                if !stringValue.isEmpty {
+                    components.append(stringValue)
+                }
+            }
+        }
+        return components
+    }
+    
+    /// Redraws the NSTextAttachments in the NSTextContainer to fit the current bounds
+    @objc
+    private func redrawTextAttachments() {
+        
+        guard images.count > 0 else { return }
+        let range = NSRange(location: 0, length: attributedText.length)
+        attributedText.enumerateAttribute(.attachment, in: range, options: [], using: { value, range, stop -> Void in
+            if let attachment = value as? NSTextAttachment, let image = attachment.image {
+                
+                // Calculates a new width/height ratio to fit the image in the current frame
+                let newWidth = frame.width - 2 * (textContainerInset.left + textContainerInset.right)
+                let ratio = image.size.height / image.size.width
+                attachment.bounds.size = CGSize(width: newWidth, height: ratio * newWidth)
+            }
+        })
+        layoutManager.invalidateLayout(forCharacterRange: range, actualCharacterRange: nil)
+    }
+    
 }